+++ app/code/local/PayU/Account/controllers/PaymentController.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- *	ver. 1.8
+ *	ver. 1.8.1
  *	PayU Standard Payment Controller
  *
  *	@copyright  Copyright (c) 2011-2012 PayU