+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## 1.8
+## 1.8.1
+
+* Fixed GranTotal Amount to SubTotal
+* Fixed updatePaymentStatusCompleted for Self-Returns
+
+## 1.8.0
 
 * SDK 1.9.2 compatible
 * Fixed PayU order cancelling
@@ -7,7 +12,7 @@
 * Fixed updating customer data
 * Changed order number in  PayU description
 
-## 1.7
+## 1.7.0
 
 * Fixed problem with accepting and cancelling order in PayU [Issue #6](https://github.com/PayU/plugin_magento_160/issues/6)
 * Removed PayU.php file